Corrected vertex error

+++ PoI.h
@@ -6,16 +6,16 @@ using namespace std;
 
 class PoI{
     int id;
-    float x;
-    float y;
+    double x;
+    double y;
     string tipo;
 public:
-    PoI(int id, float x, float y) { this->id=id; this->x=x; this->y=y;}
+    PoI(int id, double x, double y) { this->id=id; this->x=x; this->y=y;}
     int getId(){return id;}
-    float getX() {  return x;}
-    float getY() {  return y;}
-    void setX(float x) { this->x=x;}
-    void setY(float y) { this->y=y;}
+    double getX() {  return x;}
+    double getY() {  return y;}
+    void setX(double x) { this->x=x;}
+    void setY(double y) { this->y=y;}
     void setID(int id) {this->id=id;}
     void setType(string type) {tipo = type;}
     bool operator==(PoI p){
